fix some bugs

+++ main.py
@@ -168,10 +168,6 @@ def logout():
 def answer():
     user_message = request.form['prompt']
 
-    if not current_user.is_authenticated:
-        # If the user is not authenticated, return an appropriate response
-        return jsonify({"error": "You must be logged in to use this feature. Please log in or register."}), 401
-
     # Get conversations only for the current user
     user_conversations = Memory.query.filter_by(owner_id=current_user.id).all()
 
@@ -248,6 +244,10 @@ def answer():
     print(f'User Input: {user_message} 😎')
     print(f'LLM Response:\n{assistant_reply} 😝\n')
 
+    #if not current_user.is_authenticated:
+    #    # If the user is not authenticated, return an appropriate response
+    #    return jsonify({"error": "You must be logged in to use this feature. Please log in or register."}), 401
+
     # Return the response as JSON, including both text and the path to the audio file
     return jsonify({
         "answer_text": assistant_reply,
@@ -269,11 +269,6 @@ def answer():
 
 @app.route('/show-history')
 def show_story():
-    if not current_user.is_authenticated:
-        # If the user is not authenticated, return an appropriate response
-        # return jsonify({"error": "You must be logged in to use this feature. Please register then log in."}), 401
-        return (f'<h1 style="color:red; text-align:center; font-size:3.7rem;">First Get Registered<br>'
-                f'Then Log Into<br>¡!¡ 😎 ¡!¡</h1>')
 
     summary_conversation = memory_summary.load_memory_variables({})
     memory_load = memory.load_memory_variables({})
@@ -289,6 +284,7 @@ def show_story():
 
 @app.route("/private-conversations")
 def get_private_conversations():
+
     if not current_user.is_authenticated:
         # If the user is not authenticated, return an appropriate response
         # return jsonify({"error": "You must be logged in to use this feature. Please register then log in."}), 401
@@ -322,6 +318,12 @@ def get_private_conversations():
 
 @app.route('/delete-conversation', methods=['GET', 'POST'])
 def delete_conversation():
+    if not current_user.is_authenticated:
+        # If the user is not authenticated, return an appropriate response
+        # return jsonify({"error": "You must be logged in to use this feature. Please register then log in."}), 401
+        return (f'<h1 style="color:red; text-align:center; font-size:3.7rem;">First Get Registered<br>'
+                f'Then Log Into<br>¡!¡ 😎 ¡!¡</h1>')
+
     form = DeleteForm()
 
     if form.validate_on_submit():
